code consistency
- Changed constant default config to defaultConfig for consistency, see constant naming in import-path-resolver.ts.
- Moved prepareConfig to config.ts.
- Changed index.ts->prepareConfig import to use ./helpers for consistency.
- Removed prepareConfig default value because it is unnecessary.
- Added jsdoc to prepareConfig, loadConfig, resolveTsConfigExtendsPath.
- Added output to loadConfig for error logging.
