+++ src/index.ts
@@ -1,6 +1,6 @@
 import { watch } from 'chokidar';
 import { sync } from 'globby';
-import { replaceAlias, replaceAliasString } from './helpers';
+import { replaceAlias, replaceAliasString, prepareConfig } from './helpers';
 import {
   ReplaceTscAliasPathsOptions,
   IConfig,
@@ -8,7 +8,6 @@ import {
   IProjectConfig,
   AliasReplacerArguments
 } from './interfaces';
-import { prepareConfig } from './helpers/config-preparer';
 
 // export interfaces for api use.
 export {
@@ -19,7 +18,7 @@ export {
   IProjectConfig
 };
 
-const DEFAULT_CONFIG = {
+const defaultConfig = {
   watch: false,
   verbose: false,
   declarationDir: undefined,
@@ -28,7 +27,7 @@ const DEFAULT_CONFIG = {
 };
 
 export async function replaceTscAliasPaths(
-  options: ReplaceTscAliasPathsOptions = { ...DEFAULT_CONFIG }
+  options: ReplaceTscAliasPathsOptions = { ...defaultConfig }
 ) {
   const config = await prepareConfig(options);
   const output = config.output;
@@ -86,7 +85,7 @@ export type SingleFileReplacer = (input: {
 }) => string;
 
 export async function prepareSingleFileReplaceTscAliasPaths(
-  options: ReplaceTscAliasPathsOptions = { ...DEFAULT_CONFIG }
+  options: ReplaceTscAliasPathsOptions = { ...defaultConfig }
 ): Promise<SingleFileReplacer> {
   const config = await prepareConfig(options);
